refactor java method call
